Finish Dijkstra Benchmarking and reduce memory usage

+++ benchmark/inc/graph.hpp
@@ -5,8 +5,8 @@
 
 // Edge Struct
 struct Edge {
-    uint32_t outgoing;
-    uint32_t weight;
+    uint16_t outgoing;
+    uint16_t weight;
 };
 
 // Abstract base class for graphs
@@ -14,16 +14,16 @@ class Graph
 {
 private:
 protected:
-    const uint32_t v_;
+    const uint16_t v_;
 
   public:
-    Graph(uint32_t v);
+    Graph(uint16_t v);
     virtual ~Graph() = default;
 
 
-    virtual void addEdge(uint32_t start, uint32_t end, uint32_t weight) = 0;
-    virtual std::vector<Edge> edgesFromStart(uint32_t start) const = 0;
-    uint32_t getNumVertices() const;
+    virtual void addEdge(uint16_t start, uint16_t end, uint16_t weight) = 0;
+    virtual std::vector<Edge> edgesFromStart(uint16_t start) const = 0;
+    uint16_t getNumVertices() const;
 };
 
 class GraphAdjList : public Graph
@@ -32,21 +32,21 @@ private:
     std::vector<std::vector<Edge>> adjList_;
 
   public:
-    GraphAdjList(uint32_t v);
+    GraphAdjList(uint16_t v);
     ~GraphAdjList();
 
-    void addEdge(uint32_t start, uint32_t end, uint32_t weight) override;
-    std::vector<Edge> edgesFromStart(uint32_t start) const override;
+    void addEdge(uint16_t start, uint16_t end, uint16_t weight) override;
+    std::vector<Edge> edgesFromStart(uint16_t start) const override;
 };
 
 class GraphAdjMatrix : public Graph {
-    std::vector<uint32_t> adjMatrix_;
+    std::vector<uint16_t> adjMatrix_;
 
   public:
-        GraphAdjMatrix(uint32_t v);
+        GraphAdjMatrix(uint16_t v);
         ~GraphAdjMatrix();
 
-    void addEdge(uint32_t start, uint32_t end, uint32_t weight) override;
-    std::vector<Edge> edgesFromStart(uint32_t start) const override;
+    void addEdge(uint16_t start, uint16_t end, uint16_t weight) override;
+    std::vector<Edge> edgesFromStart(uint16_t start) const override;
 
 };
